src/main/java/uk/gov/di/mobile/wallet/cri/healthcheck/HealthCheckResource.java
package uk.gov.di.mobile.wallet.cri.healthcheck;

import jakarta.ws.rs.GET;
import jakarta.ws.rs.Path;
import jakarta.ws.rs.Produces;
import jakarta.ws.rs.core.Response;

import static jakarta.ws.rs.core.MediaType.APPLICATION_JSON;
import static jakarta.ws.rs.core.Response.Status.OK;

@Path("/")
public class HealthCheckResource {

    @GET
    @Path("healthcheck")
    @Produces(APPLICATION_JSON)
    public Response healthCheck() {

        return Response.status(OK).entity("Service is available").build();
    }
}
